-Se mejoro la carga del dashboard, se hace la carga de forma independiente.
